newsspark/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
langgraph
langchain
langchain-groq
langchain-chroma
langchain-huggingface
langchain-community
langsmith
langchain-core
chromadb
sentence-transformers
motor
aiosqlite
httpx
Pillow
apscheduler
newsapi-python
pydantic
jinja2
python-multipart
itsdangerous
feedparser
pygooglenews
duckduckgo-search
websockets
